ref(nav): move overlay
